Moved seed helpers to own file for easier testing.
Also added already some small changes.
